Update Timer when settings are changed
The timer was not updating before when the settings were changed. This
was because the default setting for the PrayTime model was being reset
to an unparsable date when the settings were being modified. So the
default setting for the PrayTime model is set every time the
getTimerDifference is calculated.